Tracebox 0.1
This is tracebox restarted from scratch. Tracebox is now fully build with LUA
support. It uses libcrafter from (https://github.com/pellegre/libcrafter)
which is a scapy-like packet generator library in C++.

Bindings from a subset of libcrafter is available in LUA for a much more flexible
probe generation. For example:
$ tracebox -i en1 -p "ip{}/RR({'2.2.2.2','4.4.4.4'})/tcp{dport=80}/SACK({42,43})" 8.8.8.8
This command line performs a tracebox for a IPv4/TCP probe to destination 8.8.8.8:80.
The probes allso contains an IP RouteRecord option with intermediate router 2.2.2.2 and
4.4.4.4 as well as the SACK TCP option with one block: 42-43. The example is not realistic
but gives an idea on the flexibility of tracebox 0.1. Note that the destination address
can be specified directly within the probe (ip{dst='8.8.8.8'}).

As of this commit tracebox only performs a traceroute-like operation and so does not yet detects
the changes of intermediate middleboxes.

Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>