Display the modification that occured with (-v)
A verbose mode has been added to display more information about the
modifications that occured on the fields. More specificaly when the
verbose mode is enabled (with -v, disabled by default) the original
and the modified value is printed for each modified field. Example
output:

tracebox to 81.200.198.6 (bahn.de): 64 hops max
1: ctpythagore.sri.ucl.ac.be (130.104.123.254)
2: ge.cr2.brueve.belnet.net (193.191.3.85) IP::TTL (2 -> 1) IP::CheckSum (0xc299 -> 0xef6d)
3: 10ge.cr1.brueve.belnet.net (193.191.16.21) IP::TTL (3 -> 1) IP::CheckSum (0xc199 -> 0xef6d)
4: amdgw1.arcor-ip.net (195.69.145.123) IP::TTL (4 -> 1) IP::CheckSum (0xc099 -> 0xef6d)
5: bln-145-254-5-118.arcor-ip.net (145.254.5.118) IP::TTL (5 -> 1) IP::CheckSum (0xbf99 -> 0xef6d)
6: 145.253.27.186 (145.253.27.186) IP::TTL (6 -> 1) IP::CheckSum (0xbe99 -> 0xef6d)
7: 81.200.192.224 (81.200.192.224) IP::TTL (7 -> 1) IP::CheckSum (0xbd99 -> 0xef6d)
8: 81.200.194.232 (81.200.194.232) IP::TTL (8 -> 1) IP::CheckSum (0xbc99 -> 0xef6d)
9: 81.200.197.11 (81.200.197.11) IP::TTL (9 -> 1) IP::CheckSum (0xbb99 -> 0xef6d)
10: 81.200.198.6 (81.200.198.6) TCP::CheckSum (0xcfeb -> 0x61f4) IP::TTL (10 -> 1) IP::CheckSum (0xba99 -> 0xef6d) TCPOptionMaxSegSize::MaxSegSize (9000 -> 1460)

The above example shows that the MSS value has been changed from 9000 bytes
to a value of 1460 bytes by a middleboxe close to the destination.

Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>