Fix: Prevent lua GC from collecting in-use objects
Previously, multiple lua object would be created and all of
them could happen to reference the same underlying cpp objects
e.g. in l_Packet_layer
When the GC would run on of them, it would then deallocate the
underlying shared object.

This patch adds an intermediate layer to control which lua object
has the ownership of a cpp object, represented by the l_weak_ref
type.

Object which require others to be kept alive to be validated are
represented by l_strong_ref types.

Typical example is the relationship between Layers and Packet objects:
The base Packet object cannot be destroyed as long as there are
copies of some of its layers. Upon deletion of one of the layers,
this should not automatically free up the layer, but rather allow
the packet to be GCed if needed.