Default IPv4 ID field to a random value
See #58.

FreeBSD/macOS will not accept a 0-value (or interpret it as
'please autofill'?), causing it to be randomly set by the kernel, hence
tracebox has always set it to a random value on those platforms.
Made this the same behavior across platforms (and a 0 value can still be
explicitely set on Linux if passed as parameter).

Made the IP ID field R/W to test this.

Signed-off-by: Olivier Tilmans <olivier.tilmans@uclouvain.be>